First Commit Many Things
-Makes it so the Tranq and Poison Particle doesn't show up on invis players
-The Medkit's Poison Duraction and The Tranq gun's Tranq Slowness, Duraction, and Effect are now changable by their TXT files.

+++ src/game/shared/tf/tf_player_shared.h
@@ -189,7 +189,7 @@ public:
 
 	void	Burn( CTFPlayer *pPlayer, float flTime );
 	void	Poison(CTFPlayer *pPlayer, float flTime);
-	void	Tranq(CTFPlayer *pPlayer, float flTime, float flSpeed, float flEffects);
+	void	Tranq(CTFPlayer *pPlayer, float flTime, float flSpeed, int flEffects);
 
 	// Weapons.
 	CTFWeaponBase *GetActiveTFWeapon() const;
